+++ graph/grswizzle.c
@@ -104,7 +104,7 @@ copy_line_generic( unsigned char*    from,
   if (x+width < buff_width)
     width += 1;
 
-  memcpy( to, from, (unsigned int)( width * pix_bytes ) );
+  memcpy( to, from, (size_t)width * (size_t)pix_bytes );
 }
 
 
@@ -175,7 +175,7 @@ filter_rect_generic( unsigned char*   read_buff,
   read_buff  += y*read_pitch  + pix_bytes*x;
   write_buff += y*write_pitch + pix_bytes*x;
 
-  memset( temp_lines, 0, (unsigned int)(3 * pix_bytes * ( width + 2 ) ) );
+  memset( temp_lines, 0, (size_t)(3 * pix_bytes * ( width + 2 ) ) );
 
   lines[0] = (unsigned char*) temp_lines;
   lines[1] = lines[0] + pix_bytes*(width+2);
@@ -218,7 +218,7 @@ filter_rect_generic( unsigned char*   read_buff,
 
   /* process last line */
   if (y+height == buff_height)
-    memset( lines[2], 0, (unsigned int)( ( width + 2 ) * pix_bytes ) );
+    memset( lines[2], 0, (size_t)( ( width + 2 ) * pix_bytes ) );
   else
     copy_line_generic( read_buff + read_pitch, lines[2],
                        x, width, buff_width, pix_bytes );
@@ -549,7 +549,7 @@ gr_swizzle_generic( unsigned char*    read_buff,
 {
   unsigned char*  temp_lines;
   unsigned char   temp_local[ 2048 ];
-  unsigned int    temp_size;
+  size_t          temp_size;
 
   if ( height <= 0 || width <= 0 )
     return;
@@ -564,7 +564,7 @@ gr_swizzle_generic( unsigned char*    read_buff,
   * working 'lines', each of them having width+2 pixels. the first
   * and last pixels being always 0
   */
-  temp_size = (unsigned int)( ( width + 2 ) * 3 * pixbytes );
+  temp_size = (size_t)( ( width + 2 ) * 3 * pixbytes );
   if ( temp_size <= sizeof ( temp_local ) )
   {
     /* try to use stack allocation, which is a lot faster than malloc */
